contrib: remove "hooks" directory
The "hooks" directory contains a handful of example hooks. Most of these
hooks are highly specific and haven't really received any updates over
the last couple of years, except for some global cleanups. The multimail
hook has also been removed in 96455440e4c (multimail: stop shipping a
copy, 2021-06-10) in favor of its upstream project [1].

Remove those hooks. If we want to provide examples for how to use Git
hooks we should do that as part of our documentation, for example in
githooks(5).

[1]: https://github.com/git-multimail/git-multimail

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>