contrib: remove "persistent-https" remote helper
The "persistent-https" remote helper supposedly speeds up SSL operations
by running a daemon that keeps a connection open to a remote server. It
is effectively unmaintained nowadays: the last time it received an
update was in 6f1e96115d2 (contrib/persistent-https: use Git version for
build label, 2016-07-20) and its parent commits to make it compile with
Go 1.7+.

This Go toolchain is somewhat dated by now though and unsupported. The
oldest still-supported toolchain is Go 1.23, which was released in
August 2024. It is not possible to compile the remote helper with that
Go version anymore:

    $ go version
    go version go1.23.8 linux/amd64
    $ make
    case $(go version) in \
    "go version go"1.[0-5].*) EQ=" " ;; *) EQ="=" ;; esac && \
    go build -o git-remote-persistent-https \
            -ldflags "-X main._BUILD_EMBED_LABEL${EQ}GIT_VERSION=2.49.0.943.g965a70ebf62"
    go: cannot find main module, but found .git/config in /home/pks/Development/git
            to create a module there, run:
            cd ../.. && go mod init
    make: *** [Makefile:31: git-remote-persistent-https] Error 1

The problem is that modern Go toolchains require a "go.mod" file, but we
don't have any such files. This requirement exists since quite a while
already, so it's clear that nobody has tried to use this remote helper
anytime recent.

Remove the remote helper.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>